load both models at startup
