+++ segmatch/CMakeLists.txt
@@ -13,7 +13,8 @@ cs_add_library(${PROJECT_NAME}
   src/descriptors/eigenvalue_based.cpp
   src/descriptors/ensemble_shape_functions.cpp
   src/descriptors/tf_interface.cpp
-  src/descriptors/tf_interface.py
+  src/descriptors/tf_cnn_interface.py
+  src/descriptors/tf_sem_interface.py
   src/dynamic_voxel_grid.cpp
   src/features.cpp
   src/local_map.cpp
