+++ segmatch/src/descriptors/tf_cnn_interface.py
@@ -14,65 +14,7 @@ import tensorflow as tf
 import copy
 
 
-class TensorflowInterface:
-    def sem_input_callback(self, msg):
-
-        input_tensor_name = msg.input_tensor_name
-        output_tensor_name = msg.output_tensor_name
-        message_id = msg.timestamp
-
-        inputs = empty([msg.inputs.layout.dim[0].size,
-                        msg.inputs.layout.dim[1].size])
-
-        for i in range(0, msg.inputs.layout.dim[0].size):
-            for j in range(0, msg.inputs.layout.dim[1].size):
-                inputs[i][j] = msg.inputs.data[msg.inputs.layout.dim[1].stride * i + j]
-
-        self.tf_interface_semantics(inputs, input_tensor_name,
-                                    output_tensor_name, message_id)
-
-    def tf_interface_semantics(self, inputs, input_tensor_name,
-                               output_tensor_name, message_id):
-        # cnn features
-        tf.get_default_graph()
-
-        # restore variable names from previous session
-        saver = tf.train.import_meta_graph(
-            os.path.join(self.semantics_model_path, "model.ckpt.meta")
-        )
-
-        # get key tensorflow variables -- TODO Marius: load during setup
-        sem_graph = tf.compat.v1.get_default_graph()
-        sem_input = sem_graph.get_tensor_by_name(input_tensor_name + ':0')
-        sem_output = sem_graph.get_tensor_by_name(output_tensor_name + ':0')
-
-        with tf.Session() as sess:
-            saver.restore(sess, tf.train.latest_checkpoint(
-                self.semantics_model_path))
-
-            semantics = sess.run(sem_output, feed_dict={
-                sem_input: inputs},)
-
-        self.publish_sem_output(message_id, semantics)
-
-    def publish_sem_output(self, message_id, semantics):
-        out_msg = sem_output_msg()
-        layout_dim = MultiArrayDimension()
-        out_msg.timestamp = message_id
-
-        layout_dim.size = semantics.shape[0]
-        layout_dim.stride = semantics.shape[0] * semantics.shape[1]
-        out_msg.semantics.layout.dim.append(copy.copy(layout_dim))
-
-        layout_dim.size = semantics.shape[1]
-        layout_dim.stride = semantics.shape[1]
-        out_msg.semantics.layout.dim.append(copy.copy(layout_dim))
-
-        for i in range(0, out_msg.semantics.layout.dim[0].size):
-            for j in range(0, out_msg.semantics.layout.dim[1].size):
-                out_msg.semantics.data.append(semantics[i][j])
-
-        self.sem_output_publisher.publish(out_msg)
+class TensorflowCNNInterface:
 
     def cnn_input_callback(self, msg):
 
@@ -172,23 +114,16 @@ class TensorflowInterface:
         rospy.init_node('listener', anonymous=True)
         rospy.Subscriber('tf_interface_topic/cnn_input_topic',
                          cnn_input_msg, self.cnn_input_callback)
-        rospy.Subscriber('tf_interface_topic/sem_input_topic',
-                         sem_input_msg, self.sem_input_callback)
 
         self.cnn_output_publisher = rospy.Publisher(
             'tf_interface_topic/cnn_output_topic', cnn_output_msg, queue_size=50)
-        self.sem_output_publisher = rospy.Publisher(
-            'tf_interface_topic/sem_output_topic', sem_output_msg, queue_size=50)
 
         self.cnn_model_path = rospy.get_param(
             '/SegMapper/SegMatchWorker/SegMatch/Descriptors/cnn_model_path')
-        self.semantics_model_path = rospy.get_param(
-            '/SegMapper/SegMatchWorker/SegMatch/Descriptors/semantics_nn_path')
-
 
         # CNN features
-        tf.get_default_graph()
-        self.cnn_saver = tf.train.import_meta_graph(
+        tf.compat.v1.get_default_graph()
+        self.cnn_saver = tf.compat.v1.train.import_meta_graph(
             os.path.join(self.cnn_model_path, "model.ckpt.meta")
         )
         cnn_graph = tf.compat.v1.get_default_graph()
@@ -198,13 +133,12 @@ class TensorflowInterface:
             'OutputScope/descriptor_read:0')
         self.cnn_reconstruction = cnn_graph.get_tensor_by_name(
             'ReconstructionScopeAE/ae_reconstruction_read:0')
-        self.cnn_sess = tf.Session()
+        self.cnn_sess = tf.compat.v1.Session()
         self.cnn_saver.restore(self.cnn_sess, tf.train.latest_checkpoint(
             self.cnn_model_path))
-
         rospy.spin()
 
 
 if __name__ == '__main__':
-    interface = TensorflowInterface()
+    interface = TensorflowCNNInterface()
     interface.setup()
